Physics engine-ify the stick-ball collision
