+++ pooltool/physics/resolve/stick_ball/instantaneous_point/README.md
@@ -1,0 +1,1 @@
+This collision assumes the stick-ball interaction is instantaneous and point-like
